Move data hover code out of geomap code and to visualization/data-hover folder
