+++ public/app/features/visualization/data-hover/ComplexDataHoverView.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { CloseButton } from 'app/core/components/CloseButton/CloseButton';
 
-import { GeomapLayerHover } from '../event';
+import { GeomapLayerHover } from '../../../plugins/panel/geomap/event';
 
 import { DataHoverRows } from './DataHoverRows';
 import { DataHoverTabs } from './DataHoverTabs';
