+++ public/app/plugins/panel/geomap/GeomapTooltip.tsx
@@ -3,8 +3,8 @@ import { useOverlay } from '@react-aria/overlays';
 import React, { createRef } from 'react';
 
 import { Portal, VizTooltipContainer } from '@grafana/ui';
+import { ComplexDataHoverView } from 'app/features/visualization/data-hover/ComplexDataHoverView';
 
-import { ComplexDataHoverView } from './components/ComplexDataHoverView';
 import { GeomapHoverPayload } from './event';
 
 interface Props {
